+++ src/Javis.jl
@@ -265,12 +265,9 @@ function render(
         path, ext = splitext(pathname)
     end
     render_mp4 = ext == ".mp4"
-    codec_props = [:priv_data => ("crf" => "22", "preset" => "medium")]
-    if render_mp4
-        video_io = Base.open("temp.stream", "w")
-    end
-    video_encoder = nothing
-    # if we render a gif and the user hasn't set a tempdirectory => create one
+    codec_props = (crf = 23, preset = "medium")
+    video_io = nothing
+    # if we render a gif and the user hasn't set a tempdirectory
     if !render_mp4 && isempty(tempdirectory)
         tempdirectory = mktempdir()
     end
@@ -289,13 +286,14 @@ function render(
         end
         if render_mp4
             if frame == first(frames)
-                video_encoder = prepareencoder(
+                video_io = open_video_out(
+                    pathname,
                     frame_image,
                     framerate = framerate,
-                    AVCodecContextProperties = codec_props,
+                    encoder_options = codec_props,
                 )
             end
-            appendencode!(video_encoder, video_io, frame_image, filecounter)
+            write(video_io, frame_image)
         end
         filecounter += 1
     end
@@ -311,9 +309,7 @@ function render(
                $(tempdirectory)/palette.png -lavfi paletteuse -y $pathname`,
         )
     elseif ext == ".mp4"
-        finishencode!(video_encoder, video_io)
-        close(video_io)
-        mux("temp.stream", pathname, framerate; silent = true)
+        close_video_out!(video_io)
     else
         @error "Currently, only gif and mp4 creation is supported. Not a $ext."
     end
@@ -526,9 +522,10 @@ function get_javis_frame(video, objects, frame; layers = Layer[])
 
     # check if any layers have been defined
     if !isempty(layers)
-
+        starting_positions = Point[]
         # render each layer's objects and store the layer's Drawing as an image matrix
         for layer in layers
+            push!(starting_positions, layer.position)
             if isempty(Javis.CURRENT_LAYER)
                 push!(Javis.CURRENT_LAYER, layer)
             else
@@ -550,6 +547,11 @@ function get_javis_frame(video, objects, frame; layers = Layer[])
         end
 
         img_layers = place_layers(video, layers, frame)
+
+        # resetting layer positions to originals
+        for (position, layer) in zip(starting_positions, layers)
+            layer.position = position
+        end
     end
 
     empty!(CURRENT_LAYER)